Move watershed from morphology to segmentation
