+++ skimage/segmentation/heap_general.pxi
