+++ skimage/segmentation/heap_watershed.pxi
