+++ skimage/segmentation/setup.py
@@ -11,9 +11,12 @@ def configuration(parent_package='', top_path=None):
 
     config = Configuration('segmentation', parent_package, top_path)
 
-    cython(['_felzenszwalb_cy.pyx',
+    cython(['_watershed.pyx',
+            '_felzenszwalb_cy.pyx',
             '_quickshift_cy.pyx',
             '_slic.pyx'], working_path=base_path)
+    config.add_extension('_watershed', sources=['_watershed.c'],
+                         include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_felzenszwalb_cy', sources=['_felzenszwalb_cy.c'],
                          include_dirs=[get_numpy_include_dirs()])
     config.add_extension('_quickshift_cy', sources=['_quickshift_cy.c'],
