+++ skimage/segmentation/watershed.py
